Refactoring and simplification
