+++ packages/ui-plugins/live-activities/src/components/live-activity/activity-info.scss
