+++ packages/ui-plugins/live-activities/src/components/validation/PushCredentialsStatusDetails.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, Button, Flex, Heading } from '@adobe/react-spectrum';
-import { getHealthIcon } from '../utils/utils';
+import { getHealthIcon } from '../../utils/utils';
 
 interface ShouldMatch {
   app: string;
