+++ packages/ui-plugins/live-activities/src/components/validation/client-info.module.scss
