+++ packages/ui-plugins/live-activities/src/components/validation/client-validation-widget.tsx
@@ -1,10 +1,17 @@
 import React, { useMemo } from 'react';
-import { View, Button, Flex, Heading, Text, Link, ActionButton, Tooltip } from '@adobe/react-spectrum';
-import Copy from '@spectrum-icons/workflow/Copy';
-import Card from './card/card';
-import { getHealthIcon } from '../utils/utils';
-import useClientValidationStatus from '../hooks/useClientValidationStatus';
-import { useECID, useSelectedClientPushToken } from '../hooks/useClientInfo';
+import {
+  View,
+  Button,
+  Flex,
+  Heading,
+  Text,
+  Link,
+  ProgressCircle
+} from '@adobe/react-spectrum';
+import Card from '../atoms/card';
+import { CopyableValue, getHealthIcon, renderValue } from '../../utils/utils';
+import useClientValidationStatus from '../../hooks/useClientValidationStatus';
+import { useECID, useSelectedClientPushToken } from '../../hooks/useClientInfo';
 
 const MSG = {
   ecid: 'Ecid',
@@ -14,7 +21,7 @@ const MSG = {
   token: 'Push Token',
   viewCatalog: 'View Extension Catalog',
   viewInstalled: 'View Installed Extensions',
-  viewPublishing: 'View Publishing Flow',
+  viewPublishing: 'View Publishing Flow'
 };
 
 const STATUS_MESSAGE = {
@@ -23,15 +30,17 @@ const STATUS_MESSAGE = {
   'messaging-not-installed': 'Messaging Extension Not Initialized',
   'no-ecid': 'No ECID Detected',
   'no-token': 'Push Token Not Captured',
-  valid: 'Device Configured and Push Token Detected',
+  valid: 'Device Configured and Push Token Detected'
 };
 
 const STATUS_DETAILS = {
-  'no-ecid': "The ECID is automatically generated by the SDK. Missing this event generally means that either the AEP SDK or the Assurance SDK hasn't been initialized correctly.",
+  'no-ecid':
+    "The ECID is automatically generated by the SDK. Missing this event generally means that either the AEP SDK or the Assurance SDK hasn't been initialized correctly."
 };
 
 const MSG_NOT_INSTALLED = {
-  para1: "It appears you haven't properly installed and initiated the Messaging extension in your application. Make you've done the following steps:",
+  para1:
+    "It appears you haven't properly installed and initiated the Messaging extension in your application. Make you've done the following steps:",
   androidAddStep: 'Add the Messaging extension dependency to your ',
   androidRegisterStepPre: 'Include the API call to register the Messaging extension in your',
   androidRegisterStepPost: 'method:',
@@ -49,69 +58,37 @@ const MSG_NO_TOKEN = {
   iosSetup: 'Article: Asking Permission to Use Notifications',
   missingSDK: 'Missing SDK call in',
   para1: "The Push ID wasn't detected on the device. This could be for several reasons:",
-  userDeclined: 'User declined receiving push notifications',
+  userDeclined: 'User declined receiving push notifications'
 };
 
 const MSG_NO_EDGE = {
-  para1: "It appears you haven't configured your Edge extension in Launch. First, make sure you have installed and configured the extension.",
+  para1:
+    "It appears you haven't configured your Edge extension in Launch. First, make sure you have installed and configured the extension.",
   para2: 'Next, make sure the configuration has been published using the Publishing flow.',
-  para3Pre: 
-    "Finally, make sure you aren't overriding the",
+  para3Pre: "Finally, make sure you aren't overriding the",
   para3Post: 'configuration value in your app code.'
 };
 
 const MSG_NO_CONFIG = {
-  para1: "It appears you haven't configured your Messaging extension in Launch. First, make sure you have installed and configured the extension.",
+  para1:
+    "It appears you haven't configured your Messaging extension in Launch. First, make sure you have installed and configured the extension.",
   para2: 'Next, make sure the configuration has been published using the Publishing flow.',
   para3Pre: "Finally, make sure you aren't overriding the",
   para3Post: 'configuration value in your app code.'
 };
 
 const chooseStatus = status => (!status ? 'valid' : 'invalid');
+
 const chooseStatusMessage = status => {
   if (!status) return STATUS_MESSAGE.valid;
-  if ((status || '').indexOf('messaging-not-installed') !== -1) return STATUS_MESSAGE['messaging-not-installed'];
-  if ((status || '').indexOf('no-token') !== -1) return STATUS_MESSAGE['no-token'];
+  if ((status || '').indexOf('messaging-not-installed') !== -1)
+    return STATUS_MESSAGE['messaging-not-installed'];
+  if ((status || '').indexOf('no-token') !== -1) 
+    return STATUS_MESSAGE['no-token'];
   return STATUS_MESSAGE[status] || '';
 };
+
 const chooseStatusDetails = status => STATUS_DETAILS[status];
-
-function UnknownBadge() {
-  return (
-    <span
-      style={{
-        background: '#e34850',
-        color: 'white',
-        borderRadius: 4,
-        padding: '2px 12px',
-        fontWeight: 500
-      }}
-    >
-      {MSG.unknown}
-    </span>
-  );
-}
-
-function renderValue(value) {
-  if (value == null || value === 'N/A') {
-    return <UnknownBadge />;
-  }
-  return String(value);
-}
-
-function CopyableValue({ value }) {
-  if (!value) return <UnknownBadge />;
-  return (
-    <Flex alignItems="center" gap="size-100">
-      <Text>{value}</Text>
-      <Tooltip UNSAFE_style={{ zIndex: 9999 }}>
-        <ActionButton isQuiet onPress={() => navigator.clipboard.writeText(value)} aria-label="Copy">
-          <Copy size="XS" />
-        </ActionButton>
-      </Tooltip>
-    </Flex>
-  );
-}
 
 // TODO: Implement these functions import the functionality of open.app.saga.js
 const onOpenCatalog = () => {};
@@ -125,20 +102,24 @@ const DeviceWidget = ({ ...extraProps }) => {
   const ecid = useECID();
   const token = useSelectedClientPushToken();
   const status = useClientValidationStatus();
+  const isLoading = status === 'loading';
   // Data rows for ECID and Push Token
-  const dataRows = useMemo(() => [
-    {
-      label: MSG.ecid,
-      value: ecid,
-      showCopy: true
-    },
-    {
-      label: MSG.token,
-      value: token,
-      showCopy: true,
-      isLongData: true
-    }
-  ], [ecid, token]);
+  const dataRows = useMemo(
+    () => [
+      {
+        label: MSG.ecid,
+        value: ecid,
+        showCopy: true
+      },
+      {
+        label: MSG.token,
+        value: token,
+        showCopy: true,
+        isLongData: true
+      }
+    ],
+    [ecid, token]
+  );
 
   // Status details rendering
   let statusDetails: React.ReactNode = null;
@@ -149,8 +130,10 @@ const DeviceWidget = ({ ...extraProps }) => {
         {status === 'messaging-not-installed-android' ? (
           <ol>
             <li>
-              {MSG_NOT_INSTALLED.androidAddStep}<i>build.gradle</i>
-              <br /><br />
+              {MSG_NOT_INSTALLED.androidAddStep}
+              <i>build.gradle</i>
+              <br />
+              <br />
             </li>
             <li>
               {MSG_NOT_INSTALLED.importStep}
@@ -159,7 +142,8 @@ const DeviceWidget = ({ ...extraProps }) => {
               </pre>
             </li>
             <li>
-              {MSG_NOT_INSTALLED.androidRegisterStepPre} <i>onCreate()</i> {MSG_NOT_INSTALLED.androidRegisterStepPost}
+              {MSG_NOT_INSTALLED.androidRegisterStepPre} <i>onCreate()</i>{' '}
+              {MSG_NOT_INSTALLED.androidRegisterStepPost}
               <pre>
                 <code>Messaging.registerExtension();</code>
               </pre>
@@ -169,7 +153,8 @@ const DeviceWidget = ({ ...extraProps }) => {
           <ol>
             <li>
               {MSG_NOT_INSTALLED.iosAddStep}
-              <br /><br />
+              <br />
+              <br />
             </li>
             <li>
               {MSG_NOT_INSTALLED.importStep}
@@ -178,7 +163,8 @@ const DeviceWidget = ({ ...extraProps }) => {
               </pre>
             </li>
             <li>
-              {MSG_NOT_INSTALLED.iosRegisterStepPre} <i>onCreate()</i> {MSG_NOT_INSTALLED.iosRegisterStepPost}
+              {MSG_NOT_INSTALLED.iosRegisterStepPre} <i>onCreate()</i>{' '}
+              {MSG_NOT_INSTALLED.iosRegisterStepPost}
               <pre>
                 <code>AEPMessaging.registerExtension();</code>
               </pre>
@@ -299,27 +285,43 @@ const DeviceWidget = ({ ...extraProps }) => {
   const healthIconStatus = chooseStatus(status);
 
   return (
-    <Card {...extraProps}>
-      <Heading marginTop="size-0">{MSG.title}</Heading>
-      <Text>{MSG.info}</Text>
-      <Flex gap="size-100" alignItems="center" marginTop="size-200">
-        {getHealthIcon(healthIconStatus, 'L')}
-        <Heading level={4}>{statusText}</Heading>
-      </Flex>
-      {statusDetails && <View marginTop="size-200">{statusDetails}</View>}
-      <table style={{ width: '100%', borderCollapse: 'collapse', marginTop: 16 }}>
-        <tbody>
-          {dataRows.map(row => (
-            <tr key={row.label} style={{ borderBottom: '1px solid #e1e1e1' }}>
-              <td style={{ fontWeight: 500, padding: '12px 16px' }}>{row.label}</td>
-              <td style={{ padding: '12px 16px', maxWidth: row.isLongData ? 300 : undefined, wordBreak: row.isLongData ? 'break-all' : undefined }}>
-                {row.showCopy ? <CopyableValue value={row.value} /> : renderValue(row.value)}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </Card>
+    <View>
+      <Card {...extraProps}>
+        <Heading marginTop="size-0">{MSG.title}</Heading>
+        {isLoading ? (
+          <Flex justifyContent="center" alignItems="center" height="100px">
+            <ProgressCircle aria-label="Loading…" isIndeterminate />
+          </Flex>
+        ) : (
+          <>
+            <Text>{MSG.info}</Text>
+            <Flex gap="size-100" alignItems="center" marginTop="size-200">
+              {getHealthIcon(healthIconStatus, 'L')}
+              <Heading level={4}>{statusText}</Heading>
+            </Flex>
+            {statusDetails && <View marginTop="size-200">{statusDetails}</View>}
+            <table style={{ width: '100%', borderCollapse: 'collapse', marginTop: 16 }}>
+              <tbody>
+                {dataRows.map(row => (
+                  <tr key={row.label} style={{ borderBottom: '1px solid #e1e1e1' }}>
+                    <td style={{ fontWeight: 500, padding: '12px 16px' }}>{row.label}</td>
+                    <td
+                      style={{
+                        padding: '12px 16px',
+                        maxWidth: row.isLongData ? 300 : undefined,
+                        wordBreak: row.isLongData ? 'break-all' : undefined
+                      }}
+                    >
+                      {row.showCopy ? <CopyableValue value={row.value} /> : renderValue(row.value)}
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+          </>
+        )}
+      </Card>
+    </View>
   );
 };
