+++ packages/ui-plugins/live-activities/src/hooks/useLaunchLiveActivity.ts
@@ -35,7 +35,7 @@ function useLaunchLiveActivity() {
       }
     });
 
-    console.log(JSON.stringify(command));
+    // console.log(JSON.stringify(command));
     sendCommand(command);
   };
