+++ packages/ui-plugins/live-activities/src/hooks/useSelectedActivity.ts
