+++ packages/ui-plugins/live-activities/src/utils/utils.tsx
@@ -2,8 +2,10 @@ import React from 'react';
 import Checkmark from '@spectrum-icons/workflow/Checkmark';
 import Alert from '@spectrum-icons/workflow/Alert';
 import Info from '@spectrum-icons/workflow/Info';
-import { ProgressCircle } from '@adobe/react-spectrum';
+import { ActionButton, Flex, ProgressCircle, Tooltip, Text } from '@adobe/react-spectrum';
 import Help from '@spectrum-icons/workflow/Help';
+import UnknownBadge from '../components/atoms/UnknownBadge';
+import Copy from '@spectrum-icons/workflow/Copy';
 
 // Only allow valid sizes for Spectrum icons and ProgressCircle
 const allowedSizes = ['S', 'M', 'L'] as const;
@@ -54,3 +56,29 @@ export const getHealthIcon = (status, size = 'S') => {
       );
   }
 };
+
+export function renderValue(value) {
+  if (value == null || value === 'N/A') {
+    return <UnknownBadge message= {'Unknown'} />;
+  }
+  return String(value);
+}
+
+
+export function CopyableValue({ value }) {
+  if (!value) return <UnknownBadge message= {'Unknown'} />;
+  return (
+    <Flex alignItems="center" gap="size-100">
+      <Text>{value}</Text>
+      <Tooltip UNSAFE_style={{ zIndex: 9999 }}>
+        <ActionButton
+          isQuiet
+          onPress={() => navigator.clipboard.writeText(value)}
+          aria-label="Copy"
+        >
+          <Copy size="XS" />
+        </ActionButton>
+      </Tooltip>
+    </Flex>
+  );
+}
